Update cipherans.py.txt
This file was not supposed to the main solution but an earlier draft, however I mistakenly uploaded the link to this branch instead of “linuxbranch” branch which contains my final solution which was uploaded on May 2nd. I made a very minor edit on the variable “cipher_ans” to ensure this code matched my final draft.